Updates to get the project building again, fixing new warnings and deprectation issues

+++ app/src/main/java/com/geekyouup/android/ustopwatch/fragments/LapTimesFragment.java
@@ -86,16 +86,14 @@ public class LapTimesFragment extends ListFragment implements LapTimeListener {
                 @Override
                 public boolean onActionItemClicked(android.view.ActionMode actionMode, android.view.MenuItem menuItem) {
                     // Respond to clicks on the actions in the CAB
-                    switch (menuItem.getItemId()) {
-                       /* case R.id.menu_context_delete:
-                            mLapTimeRecorder.deleteLapTimes(mCheckedItems, ltf);
-                            actionMode.finish(); // Action picked, so close the CAB
-                            mCheckedItems.clear();
-                            mCheckedItems = null;
-                            return true;*/
-                        default:
-                            return false;
-                    }
+                    if(menuItem.getItemId() == R.id.menu_context_delete)
+                    {
+                        mLapTimeRecorder.deleteLapTimes(mCheckedItems, ltf);
+                        actionMode.finish(); // Action picked, so close the CAB
+                        mCheckedItems.clear();
+                        mCheckedItems = null;
+                        return true;
+                    }else return false;
                 }
 
                 @Override
